requirements.txt
allure-pytest==2.8.22
allure-python-commons==2.8.22
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
idna==2.10
iniconfig==1.1.1
packaging==20.5
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
